Merge pull request #12 from highthon/getUserByToken

+++ src/main/kotlin/com/example/highthon/global/config/jwt/env/TokenProperty.kt
@@ -7,6 +7,6 @@ import org.springframework.boot.context.properties.ConstructorBinding
 @ConstructorBinding
 data class TokenProperty(
     val secretKey: String,
-    val accessExp: Int,
-    val refreshExp: Int
+    val accessExp: Long,
+    val refreshExp: Long
 )
